requirements.txt
Flask>=3.1
flask-jwt-extended>=4.6
flask-limiter>=4.1
flask-socketio>=5.4
flask-wtf>=1.2
cryptography>=43
requests>=2.32
psycopg2-binary>=2.9
PyYAML>=6.0
flask-cors>=4.0
redis>=5.0
eventlet>=0.36
gunicorn>=22.0
argon2-cffi>=23.1
livekit-api>=1.1.0
